requirements.txt
aiohttp==3.3.2
cchardet==2.1.1
aiodns==1.1.1
backoff==1.6.0
uvloop==0.11.0
PyExecJS==1.5.1
dnspython==1.15.0
